Impute values in X by randomly selecting data points from recorded observations in Y from an external, pre-existing data source.

Cold deck imputation (CDI) is a univariate imputation technique where, for each respondent or recipient with a missing value, we use an external, pre-existing source to find a donor with similar values across a subset of categorical or numerical predictors and use it to fill the recipient's missing observation. For this reason, cold deck imputation has been used with stratification across categorical variables.

The current implementation of CDI allows the user to choose one of four selection methods: deterministic, random sampling from all possible donors, from k-nearest neighbors and random sampling using weights as probabilities. The function will iteratively impute missing values across all variables with missing observations using the selection method specified in the function arguments.

Usage

coldeck.impute(
  data,
  ext_data = NULL,
  method = "deterministic",
  k = NULL,
  seed = NULL,
  na.rm = TRUE
)

Arguments

data

a matrix or data frame containing missing values in at least one predictor

ext_data

external data source of complete cases to be used as donor values

method

selection method for imputing missing values based on donor similarity. Can be one of:

"deterministic"

Select the same donor value for multiple repetitions of the CDI.

"rand_from_all"

Select a different donor value for each repetition of the CDI.

"rand_nearest_k"

Select one random donor value from a subset of k nearest neighbors for each repetition of the CDI.

"weighted_rand"

Select one random donor through a probability-weighted choice for each repetition of the CDI.

k

number of nearest neighbors to select from when using the rand_nearest_k method.

seed

a numeric seed for reproducible results for every method except deterministic selection

na.rm

indicates removal of NA values from every row in the matrix or data frame

Value

a matrix or data frame of imputed values

Details

CDI is a valuable method when a reliable external source of data is available and can ensure more standardized imputations, particularly in large-scale studies or when historical consistency is important. However, it requires careful consideration to avoid introducing bias due to mismatches between the current dataset and the external source.

Examples

data <- gen.mcar(100,rho = c(.56,.23,.18),sigma = c(1,2,.5),n_vars = 3,na_prob = .18)
ext_data <- gen.mcar(100,rho = c(.45,.26,.21),sigma = c(1.67,2.23,.56),n_vars = 3,na_prob = 0)
coldeck.impute(data,ext_data)
#>               V1          V2           V3
#> 1    1.640846166  0.03642265 -0.125325398
#> 2   -0.219050379 -0.96719453 -0.562132515
#> 3    0.168065384  0.76172181 -0.154975693
#> 4    1.168383873  0.23644489 -0.025481243
#> 5    1.054181023 -2.39450533  0.736386047
#> 6    1.145263110  2.74788522  0.787452440
#> 7   -0.577468001 -2.02119630 -0.621353744
#> 8    2.002482730  0.03642265 -0.211412046
#> 9    0.066700871  2.56664605  0.180845510
#> 10   0.615799237  0.80812736  0.279074683
#> 11  -1.350902686 -0.11164635  0.737649960
#> 12   0.020983586 -2.06543127  0.140277567
#> 13   1.249914571  0.81243280  0.400920094
#> 14  -0.715242187 -0.92295249  0.030388305
#> 15  -0.752688968 -2.77944978 -0.605223161
#> 16  -0.938538704 -2.10293184  0.371075071
#> 17  -1.052513279 -1.22660490 -0.811816426
#> 18  -0.437159533  0.62171463 -0.762420438
#> 19   0.331179173 -2.36401054 -0.059120306
#> 20  -2.014210498  1.05599169 -0.677525563
#> 21   0.211980433  1.49077237 -0.957833818
#> 22   1.236675046  0.49227724  0.198469025
#> 23   0.615799237  2.65870225  0.202877962
#> 24   1.301175992  2.27293257  0.117561046
#> 25  -0.161465742  0.35964979  0.200271600
#> 26  -1.726730399 -0.85149973  0.250148104
#> 27   0.615799237 -0.87701395  0.026875964
#> 28   0.004495074 -1.07974790 -0.352617145
#> 29   0.703523903 -3.59171196 -0.357594956
#> 30  -0.105671334  0.03642265 -0.079029215
#> 31  -1.258648628 -0.69671108  0.019099340
#> 32   0.004495074  2.77371596 -0.294628889
#> 33   0.911391292  0.10067023 -0.001850647
#> 34   0.237430272  3.21453192 -0.599784166
#> 35   1.218108610  1.83888267  0.096350613
#> 36  -1.338774287 -1.29012369 -0.811816426
#> 37   0.660820298  2.84824598 -0.018695898
#> 38  -0.125689540 -1.77614977 -0.562132515
#> 39   0.683745522  0.87733482  0.331824132
#> 40  -0.060821955  3.90442272  0.667631682
#> 41   0.632960713  0.72735690  0.140682130
#> 42   1.335517615  4.20257845  0.403295979
#> 43   0.007290090 -2.37541934 -0.314804835
#> 44   1.017558637  0.82398215  0.405108385
#> 45  -1.188434035 -0.70400331 -0.370859081
#> 46  -0.183216236 -0.31103754 -0.763558893
#> 47   1.519217711  0.03519942 -0.122994765
#> 48   0.377387973  0.45458686  0.988925838
#> 49  -2.052222820  1.56889743  0.119637235
#> 50  -0.125689540 -0.34678307  0.474401635
#> 51  -0.161465742  0.35964979  0.184694273
#> 52   0.865779404 -2.71709878 -0.264709781
#> 53  -0.101883256  1.93341655 -0.071745113
#> 54   0.624187472 -1.35730241 -0.599784166
#> 55  -0.125689540  1.82763198 -0.562132515
#> 56   1.671054829  2.96503032  0.554514919
#> 57   0.056016733 -0.26847575  0.474401635
#> 58  -0.051981906 -1.12716402  0.394807255
#> 59  -1.753237359 -1.68969135  0.223129930
#> 60   0.099327594  0.83836300 -0.557084943
#> 61  -0.571850058  0.82314765  0.272076634
#> 62  -0.974009583 -4.49158277  1.004722974
#> 63  -0.179906231 -2.91295123 -0.341843720
#> 64   1.014943173  3.50689188  1.334308992
#> 65  -0.161465742  0.35964979 -0.576673352
#> 66  -0.427279287  0.24271455  0.503689262
#> 67   0.116637284  1.31567380  0.156830129
#> 68  -0.893207570  0.51935280  0.727813358
#> 69   0.333902942 -4.03513844 -0.752317829
#> 70   0.411429921  2.30051424  0.056710505
#> 71  -0.033036159 -0.84061777 -0.274666872
#> 72  -2.465898194 -1.37298181 -0.372246782
#> 73   2.571458146  2.39096101 -0.019134314
#> 74  -0.205299257  1.21489787 -0.401591626
#> 75   0.651193282  0.15192278  0.345154855
#> 76   0.273766491  1.16577154 -0.480638948
#> 77   0.648569822  0.50027479  0.826430528
#> 78   0.817659446 -0.89495367 -0.515773359
#> 79  -0.209793171  1.77000117  0.062347869
#> 80   0.378167772  1.65120827  0.325209383
#> 81  -0.945408831  1.79821914  0.229483261
#> 82  -1.538931792  1.06771288 -0.045972806
#> 83  -0.461038339  1.34774998  0.020349825
#> 84   2.416773354  5.98002632  0.805770981
#> 85  -0.125689540 -2.31558647 -0.562132515
#> 86  -0.463987243 -2.71177615 -0.473860205
#> 87   0.825379863  0.52782344  1.334308992
#> 88   0.510132547  0.21668528 -0.163940541
#> 89  -0.589481039  0.86975625  0.602188613
#> 90  -0.996780742  1.72086960  0.265220559
#> 91   0.144475705 -0.37860282  0.041582156
#> 92  -0.014307413  0.60203890  0.646098708
#> 93  -1.790281237 -2.38238411 -0.291297758
#> 94  -0.125689540  0.07258372 -0.149050447
#> 95   0.190230316  0.73344686 -0.018120114
#> 96   0.004495074  2.39652568 -0.512646114
#> 97  -1.055017043 -0.98059633  0.124576666
#> 98   0.476133278  1.71443846 -0.427749214
#> 99   1.378570137  2.83455811  0.072745554
#> 100  0.456236403  2.02675056  0.265958649