Generate a data frame containing NAs missing at random

Generate correlated, synthetic normal variables with user-specified probability of MAR. Specify the column length, correlation coefficient, standard deviation, number of columns and desired probability of missing values to obtain a data frame of correlated observations with missing values.

Usage

gen.mar(len, rho, sigma, n_vars, na_prob)

Arguments

len

number of rows per column

rho

desired correlation coefficient of generated variables. The length of rho must be equal to the product of n_vars and half of n_vars minus one.

sigma

desired standard deviation for each generated variable

n_vars

total number of variables to be generated. At least two variables must be provided.

na_prob

desired probability of missingness in each variable.

Value

a data frame of at least 2 columns

Examples

syn_na <- gen.mar(50,c(.25,.75,.044),c(1.1,.56,1.56),3,.47)