Calculate the geometric mean for a set of non-negative, non-zero observations.

The geometric mean is the nth root of the product of 'n' observations. It multiplies the observations in a set and can be thought of as the exponential of the arithmetic mean of logarithms. We use this definition to create our function. #' The 'geom.mean' function will calculate the mean for all numeric columns in a matrix or data frame.

Usage

geom.mean(x, ..., na.rm = FALSE)

Arguments

x

a numeric vector, matrix or data frame

...

additional parameters to be passed to 'mean()'

na.rm

indicate whether or not to remove NAs

Value

a numeric vector of length equal or greater than 1L.

Details

Because we're using logarithms, our observations must be non-negative. Additionally, by using the geometric mean and SD and variance, we'll be measuring the log-normal dispersion of a log-normal distribution

Examples

x <- rlnorm(60,1,1.4)
geom.mean(x)
#> [1] 2.935312

d <- matrix(rlnorm(60,1,1.3),ncol = 4)
sapply(d,geom.mean)
#>  [1]  1.1901508  5.6155205  4.2355163  0.5866222  0.2827114  0.2038519
#>  [7]  5.5658876  2.5982427 30.1366854  5.7303559  8.2038590 15.4051006
#> [13]  1.4177416  5.2757841  8.4100822 16.1215550  7.3261233  4.6996654
#> [19]  0.8792454  7.0176505  5.2087798  6.1998514  4.7149512  2.0909333
#> [25]  2.4597859  6.6430643  3.3977755  0.9239095  1.8641276  0.4892589
#> [31]  0.7732588  2.2524209  0.2624821  0.3026721 11.3968244  1.2845332
#> [37]  0.2448280  2.3473283 15.1407540  6.4297174  4.8249393 12.6653471
#> [43]  0.9970525  7.0188259  1.2671733  2.7209836 48.3663786  9.5856263
#> [49]  7.3773170  0.6435045  0.9781359 52.9501775  0.6561991  3.5924096
#> [55]  8.6741070 10.2203963 11.1992366  2.1988452  0.9349309  1.8231497