Calculate the geometric standardized score

Compute the geometric z score of each strictly positive value in the sample, relative to the geometric mean and standard deviation typically used in log-normally distributed populations.

Usage

geom.zscore(x, ..., na.rm = FALSE)

Arguments

x

a numeric vector, matrix or data frame

...

additional parameters to be passed to 'geom.mean()'

na.rm

indicate whether or not to remove NAs

Value

a matrix or array of z-scores of length > 1L.

Details

The formula here makes use of the change of base property of logarithms.

Examples

d <- rlnorm(60,1,1.4)
geom.zscore(d)
#>  [1]  0.51862261  0.12147132  0.28977629 -0.28651514  0.15586424  1.23500259
#>  [7]  0.59327578  0.30416092  0.10337092 -0.14465639 -1.54854552 -0.97480980
#> [13]  0.82097898  0.06613354 -0.61947156  0.49103254 -1.09880321  0.06084787
#> [19] -0.93113293 -0.73112486  0.35538523 -0.70636109 -0.22322170  0.16532036
#> [25] -1.21088048  2.14030428  0.86642047  0.15116015 -1.84200519  0.82661415
#> [31]  1.05574533  0.35558409  0.32030107 -0.62752676 -1.71990875  0.08061814
#> [37]  1.59390535  0.04784997 -1.24503735  2.75079969 -0.50217198  0.40439257
#> [43]  0.71355250 -0.47420902  0.10459941 -0.48635389 -0.09370144  1.03009607
#> [49]  0.78306200 -0.55028345  1.24269928  0.64698790 -1.81379701 -0.17941193
#> [55]  1.29491910 -3.14003345 -0.64305996  0.77651911  0.31502753 -0.98937847

x <- matrix(rlnorm(60,1,1.4), ncol = 4)
mapply(geom.zscore, as.data.frame(x))
#>                V1         V2          V3          V4
#>  [1,]  0.50708316  0.7451825  0.21109638 -0.03353509
#>  [2,] -0.61547813 -0.9504166  1.35355354 -1.65204149
#>  [3,] -0.87460714 -0.8260462 -2.09030365 -0.33897951
#>  [4,] -1.88086603  0.7785544  1.29359101  0.99178967
#>  [5,]  0.78272501  1.1952791  0.07222551 -1.20958603
#>  [6,]  0.25056258 -0.1782002  0.47116385 -0.80528383
#>  [7,]  0.27501346 -1.7497917 -0.95175484 -1.64860442
#>  [8,]  1.45812495  0.6867929 -0.42820723  1.03768755
#>  [9,] -0.98787705  0.5144891 -0.27590956 -0.19404356
#> [10,] -0.43577441  0.8820399 -0.84185400  1.51020762
#> [11,] -0.66684979 -1.4192569 -0.16914113  1.55348303
#> [12,] -0.06343104  1.2959843 -0.24748348  0.32692824
#> [13,]  1.99858330  0.4134185 -0.70819942 -0.12828313
#> [14,]  1.00487677 -1.5271480  0.22904525 -0.04807204
#> [15,] -0.75208562  0.1391189  2.08217778  0.63833298