Calculate the harmonic mean for matrices, data frames and vectors.

The harmonic mean is the reciprocal of the arithmetic mean of the reciprocals of a set of observations. It is a more robust measure of central tendency than the arithmetic mean in the presence of outliers since it tends to the smallest elements of a set of observations.

Usage

harm.mean(x, na.rm = FALSE, ...)

Arguments

x

an object containing a set of observations in a matrix, data frame or vector.

na.rm

logical by default set to FALSE. To remove NAs set argument to TRUE

...

additional arguments passed to or from other methods

Value

a numeric vector of length equal or greater than 1L

Examples

d <- rlnorm(60,1,1.4)
harm.mean(d)
#> [1] 1.495732

x <- matrix(rlnorm(60,1,1.4),ncol = 4)
sapply(x,harm.mean)
#>  [1]  0.2685039  1.7605401  6.0788895  1.3066609  2.4995595  2.6394999
#>  [7]  0.9869711  0.6755205  2.0548151  1.6722350  8.7384523 23.1463036
#> [13]  2.6732866  4.7499818  2.4630947  6.3085369  3.6909861  3.7857109
#> [19] 12.9110196  0.6822292 20.6354324  4.4315853  3.5013577  1.2406822
#> [25]  2.1610006  4.5866904  0.1511927  1.1710131  6.1742552  1.7998899
#> [31]  0.8862656  1.1163089  3.7812273  0.2756109  9.2047618  0.1304835
#> [37]  1.3435625  0.9029795  0.4732932  0.6010541  3.8578433  2.2986370
#> [43]  1.7839077  0.1054974  0.4256878  2.2618206  0.7951304  0.8909415
#> [49] 11.8148474  3.4757966  0.6304115  1.4103326  1.8241883  1.0487372
#> [55]  0.7049171  2.5281158  7.6727309  3.7344211  2.0540104  4.6783634