Hot deck imputation

Hot deck imputation (HDI) is a univariate imputation technique where, for each respondent or recipient with a missing value, we find a donor with similar values across a subset of categorical or numerical predictors and use it to fill the recipient's missing observation. For this reason, HDI has been used with stratification across categorical variables.

The current implementation of HDI allows the user to choose one of four selection methods: deterministic, random sampling from all possible donors, from k-nearest neighbors and random sampling using weights as probabilities. The function will iteratively impute missing values across all variables with missing observations using the selection method specified in the function arguments.

Usage

hotdeck.impute(
  data,
  method = "deterministic",
  k = NULL,
  seed = NULL,
  na.rm = TRUE
)

Arguments

data

a matrix or data frame containing missing values in at least one predictor

method

selection method for imputing missing values based on donor similarity. Can be one of:

"deterministic"

Select the same donor value for multiple repetitions of the CDI.

"rand_from_all"

Select a different donor value for each repetition of the CDI.

"rand_nearest_k"

Select one random donor value from a subset of k nearest neighbors for each repetition of the CDI.

"weighted_rand"

Select one random donor through a probability-weighted choice for each repetition of the CDI.

k

number of nearest neighbors to select from when using the rand_nearest_k method.

seed

a numeric seed for reproducible results for every method except deterministic selection

na.rm

indicates removal of NA values from every row in the matrix or data frame

Value

a matrix or data frame of imputed values

Examples

data <- gen.mcar(100,rho = c(.56,.23,.18),sigma = c(1,2,.5),n_vars = 3,na_prob = .18)
hotdeck.impute(data)
#>               V1          V2           V3
#> 1    0.793918367  0.34842553  0.096484456
#> 2   -0.140513958  1.30490257  0.210867432
#> 3    0.455805199  2.63357399 -0.284916489
#> 4   -1.145572907 -2.47433574  0.401348744
#> 5   -0.249650688  0.82157465  1.060543391
#> 6   -0.420298275 -1.40719171  0.651553088
#> 7    0.195664504 -1.37517092 -0.227900727
#> 8    0.357319514  1.43309765  0.301787300
#> 9   -0.123617979  0.03801310  0.015393905
#> 10  -0.766214223 -0.21304442  0.497724369
#> 11  -0.929714217 -2.00316563 -0.311322416
#> 12   0.455805199  3.28619502 -0.284916489
#> 13   1.356836852  2.62202967 -0.363966359
#> 14  -0.787135595 -1.50197115 -0.037141272
#> 15  -0.384798672  0.73347007  0.301787300
#> 16   0.330680560 -1.64207015  0.424407564
#> 17  -0.554620450 -0.84942483 -0.841088182
#> 18   0.121572315  2.46150428 -0.284916489
#> 19  -0.047596117 -0.36621160 -0.648726990
#> 20  -0.776251591  0.63926730 -1.038441499
#> 21   0.831441251  0.56349273  0.333960472
#> 22   0.846307837  0.56349273  0.428459498
#> 23   0.391720548 -0.31146293  0.253671466
#> 24   1.267996586  1.78922592 -0.447186498
#> 25  -0.506361788  1.82979125  0.747770152
#> 26  -0.464481897 -1.50347827 -0.251499344
#> 27   0.261218000  0.74985730  0.902492269
#> 28   0.630080977 -0.89300060  0.310415003
#> 29  -0.339626156 -2.06463426  0.507565624
#> 30  -0.423344808  0.54364121  0.087841477
#> 31  -0.618271528  0.95571200 -0.060193883
#> 32   1.482201891  3.21814603 -0.284916489
#> 33  -2.508166352 -1.82545193 -0.541317066
#> 34  -0.167578034  1.66061756 -0.007680438
#> 35   0.038212877 -0.66046752  0.128258867
#> 36  -1.059609603  0.07184500  0.497724369
#> 37   0.385425895  0.77902962 -0.648320996
#> 38  -1.967087684 -3.94148293 -0.169528772
#> 39   0.954968861 -0.30749269  1.060543391
#> 40  -1.663360585 -1.78002689 -0.258879342
#> 41  -2.202734880 -1.82545193 -0.953870127
#> 42  -0.763563826 -1.38767763 -0.037141272
#> 43   0.162080394 -1.46964122  0.669444315
#> 44  -0.651567165 -4.19433225 -0.180146756
#> 45  -1.605718058  0.07820747 -0.410978336
#> 46   0.333204975 -0.11607128  1.032321689
#> 47  -1.058900921 -0.27601757 -0.632560961
#> 48  -0.085849546 -2.65767137 -0.567192658
#> 49   0.497932993 -1.51907416  0.310415003
#> 50   1.633989657  1.64559315 -0.447186498
#> 51   0.479451881 -1.80402503  0.072651573
#> 52   1.714762992 -1.36519066 -0.406395918
#> 53   0.453160034  1.80594524  0.641476290
#> 54  -0.003241127  1.48909711  0.301787300
#> 55  -2.256534856  0.09721225  0.069183803
#> 56  -1.224658552 -1.82241245 -0.258879342
#> 57  -0.318962624 -1.50347827 -0.277072134
#> 58   0.112831695  1.32205763  0.301787300
#> 59   0.391720548 -0.41039171  0.158517037
#> 60   0.543648621 -0.83179666  0.424102832
#> 61  -1.063811352 -2.47433574  0.625606034
#> 62  -0.274129717 -2.06687812 -0.311322416
#> 63   0.217006032  0.96611246 -0.454732254
#> 64  -0.359385718  0.54364121  0.025292829
#> 65   0.112831695  0.75018483  1.492182341
#> 66  -0.670748026 -2.14789355 -0.311322416
#> 67  -0.846555519  0.19240668  0.544215939
#> 68  -0.081054893  0.78887164 -0.110969235
#> 69  -0.047049347  1.66061756 -0.243268640
#> 70  -1.948787086 -2.24791408 -0.006655192
#> 71  -0.673668581  0.95571200 -0.148162846
#> 72  -1.489644085 -3.14894390  0.598354026
#> 73  -1.605718058  0.09721225 -0.002727320
#> 74  -0.493883101 -1.23780127  0.167089512
#> 75  -0.160798368 -1.36183585  0.651553088
#> 76   0.283600226 -0.02328607  1.032321689
#> 77   1.091262650  0.37279281  0.514197397
#> 78   0.444400297  0.15437507 -0.648320996
#> 79   0.391720548 -0.35860837  0.295168952
#> 80  -0.526310288 -1.05611479 -0.841088182
#> 81  -0.307840173 -4.55672324 -1.497238031
#> 82   1.085168884  0.37279281  0.735879158
#> 83   0.001207184  1.66061756 -0.225530901
#> 84  -1.680244716 -2.16233503 -0.006655192
#> 85  -0.846555519  0.19240668  0.497724369
#> 86   0.455805199  1.93376881 -0.117663038
#> 87  -0.359385718  0.38580138 -0.115826706
#> 88   0.333444133  1.46051420  0.228706569
#> 89   0.014222696 -2.16233503 -0.567192658
#> 90  -0.496357607 -0.54902799 -0.432970332
#> 91  -0.350786392  2.30811671  0.364921554
#> 92   0.391720548 -0.35742990  0.295168952
#> 93   0.209578829  1.30490257 -0.101981868
#> 94   1.234670140  1.62043185 -0.447186498
#> 95  -0.199819784  0.69916286  0.301787300
#> 96  -0.923208042 -1.48474289  0.578647932
#> 97   0.165903102 -0.88841087  0.669444315
#> 98  -0.846555519  0.19240668  0.507099929
#> 99   0.455805199  2.90014909  0.034230785
#> 100 -1.058900921 -0.82641784 -0.317079321