Transform non-missing data to missing values with user-specified probability of MAR

Transform a complete case dataset according to the MAR mechanism.The MAR mechanism assumes that the probability of missingness in a variable depends on the observed data but not on the missing data itself. This function introduces missing values in selected features of a dataset, with the missing values determined by the values of the causative (target) feature.

Usage

mar.transform(input, target, features, na_rate)

Arguments

input

data to transform using MAR mechanism

target

variable to be used as causative feature

features

variables to which NA values are introduced using a causative feature

na_rate

proportion of missing values to be added to data

Value

a matrix or data frame containing NAs

Examples

set.seed(123)
data <- gen.mcar(100,rho = c(.15,.25,.12,.45,.34,.54),sigma = c(1,2,1,2),n_vars = 4, na_prob = 0)

mar.transform(data,"V1",c("V2","V3"), na_rate = .25)
#> Warning: One or more features are at risk of being entirely NA.
#>               V1          V2           V3          V4
#> 1   -0.560475647          NA           NA -0.10870782
#> 2   -0.230177489  0.43890141  1.229901703 -0.07693965
#> 3    1.558708314 -0.02018912  0.113321644 -0.34213763
#> 4    0.070508391 -0.66606849  0.512615342 -1.22758235
#> 5    0.129287735 -1.84291758 -0.446775346 -1.40156838
#> 6    1.715064987  0.42548293 -0.034399523  1.60862474
#> 7    0.460916206 -1.41377325 -0.710987188 -3.65506210
#> 8   -1.265061235          NA           NA -2.24347826
#> 9   -0.686852852          NA           NA  2.38387824
#> 10  -0.445661970          NA           NA  3.07054916
#> 11   1.224081797 -0.77045043  0.373040290  2.80692212
#> 12   0.359813827  1.31011577  0.375755980  1.98141531
#> 13   0.400771451 -3.07892452  1.154089964 -2.04864243
#> 14   0.110682716 -0.07666186 -0.474783259 -1.25249611
#> 15  -0.555841135          NA           NA -1.56643493
#> 16   1.786913137  1.13156617  2.088307907  4.21896599
#> 17   0.497850478  0.35831629 -0.292284557 -0.01978737
#> 18  -1.966617157          NA           NA -4.58910274
#> 19   0.701355902 -1.46977481 -1.088126658  1.61466054
#> 20  -0.472791408          NA           NA -0.72096614
#> 21  -1.067823706          NA           NA -1.02676922
#> 22  -0.217974915 -1.93890224  0.462581120  1.59948550
#> 23  -1.026004448          NA           NA -2.24098233
#> 24  -0.728891229          NA           NA  0.77256775
#> 25  -0.625039268          NA           NA -0.62367973
#> 26  -1.686693311          NA           NA -0.81710759
#> 27   0.837787044  0.71678293 -0.450596021  0.20298897
#> 28   0.153373118  0.20016954 -0.647011187  0.51466709
#> 29  -1.138136937          NA           NA  1.16051478
#> 30   1.253814921  0.23514186 -0.672181289  0.24816516
#> 31   0.426464221  2.98435368  2.113316899  4.32329398
#> 32  -0.295071483          NA           NA -1.84825272
#> 33   0.895125661  0.35007052  0.434175544 -0.05785329
#> 34   0.878133488 -0.57199336 -0.528137441  2.18259134
#> 35   0.821581082 -3.81355923 -0.520016385 -0.31655188
#> 36   0.688640254  2.44366660 -1.003887589 -2.89312439
#> 37   0.553917654 -2.72205347 -0.159860150 -2.47531237
#> 38  -0.061911711  1.44457797  0.450434971  0.40690741
#> 39  -0.305962664          NA           NA  2.32810853
#> 40  -0.380471001          NA           NA -1.94431513
#> 41  -0.694706979          NA           NA  0.02242443
#> 42  -0.207917278 -0.58083716 -0.558300879  0.66403287
#> 43  -1.265396352          NA           NA  1.71806274
#> 44   2.168955965 -2.34437461 -0.681243614  0.24607600
#> 45   1.207961998 -2.80444418 -0.004369181 -0.02262170
#> 46  -1.123108583          NA           NA -2.29831263
#> 47  -0.402884835          NA           NA -1.10630322
#> 48  -0.466655354          NA           NA -2.02608923
#> 49   0.779965118  4.38668614 -0.271032563 -0.13691625
#> 50  -0.083369066 -2.56994874  0.340058465 -0.64315612
#> 51   0.253318514  1.63364829 -0.233261092  1.84672281
#> 52  -0.028546755  1.51211856 -0.484975044 -3.01713255
#> 53  -0.042870457  0.64402694 -0.314754798 -0.77516374
#> 54   1.368602284 -1.58335498  0.345304508  0.92410660
#> 55  -0.225770986 -0.30393353  1.475580738 -0.24809495
#> 56   1.516470604 -0.09950470  0.270286559  1.62930176
#> 57  -1.548752804          NA           NA  0.43073881
#> 58   0.584613750 -0.56106584  0.723528484  0.80102662
#> 59   0.123854244  1.96899609  0.002864226 -3.08848203
#> 60   0.215941569 -0.67590323 -1.455975120  2.50143679
#> 61   0.379639483  2.19549401 -0.319940016  0.18161627
#> 62  -0.502323453          NA           NA -0.47917608
#> 63  -0.333207384          NA           NA -0.55051359
#> 64  -1.018575383          NA           NA  3.47416178
#> 65  -1.071791226          NA           NA  1.93673545
#> 66   0.303528641  0.68076548  1.593631500  1.42550096
#> 67   0.448209779  1.39319798  0.036727207  1.20133780
#> 68   0.053004227 -0.94071299 -1.721540313 -3.08886788
#> 69   0.922267468  1.29870877 -0.101338021  2.05494510
#> 70   2.050084686  1.34460553  0.629362247  1.44212557
#> 71  -0.491031166          NA           NA  3.71863157
#> 72  -2.309168876          NA           NA -3.67433945
#> 73   1.005738524  0.23435792  0.908706564  0.77477643
#> 74  -0.709200763          NA           NA  0.58428732
#> 75  -0.688008616          NA           NA  0.43838219
#> 76   1.025571370 -1.85952091 -0.265830082 -0.92656278
#> 77  -0.284773007          NA           NA -1.56153920
#> 78  -1.220717712          NA           NA -0.65152227
#> 79   0.181303480  0.91756035  0.494779460  0.74174666
#> 80  -0.138891362 -0.94802618 -0.049168155 -3.00198988
#> 81   0.005764186 -2.10086206 -1.695806248 -1.92834526
#> 82   0.385280401  2.61337111  0.912181480  1.94178380
#> 83  -0.370660032          NA           NA  0.05555461
#> 84   0.644376549 -1.51812793 -0.167386907 -1.67866638
#> 85  -0.220486562 -0.53335857  0.039129561  0.47385172
#> 86   0.331781964 -0.29035550  0.195791837  2.33906496
#> 87   1.096839013  2.52377700  0.580031032  2.45839981
#> 88   0.435181491  0.29811160  1.698697789  0.13672509
#> 89  -0.325931586          NA           NA -0.70501919
#> 90   1.148807618 -0.64264377  0.407662119  1.41099113
#> 91   0.993503856  0.72208931  1.393344067  1.03626986
#> 92   0.548396960 -0.47750574  1.126914994 -1.98909747
#> 93   0.238731735  0.25864634  1.172345986  2.53389439
#> 94  -0.627906076          NA           NA -2.76846310
#> 95   1.360652449 -2.18374651  2.162466427  1.33402169
#> 96  -0.600259587          NA           NA  2.83854922
#> 97   2.187332993  1.84402470  2.397805347  2.72163176
#> 98   1.532610626 -2.01444576 -1.024398036  0.80617075
#> 99  -0.235700359          NA           NA -2.39399903
#> 100 -1.026420900          NA           NA -1.05282380