Transform data to missing values completely at random

Transform a complete case dataset according to the MCAR mechanism.The MCAR mechanism assumes that the probability of missingness is the same for all cases, and therefore,the missing data are not related to the observed data. transform.mcar uses the binomial distribution to generate NA an index of possible NA values that will replace a percentage of data points in the input.

Usage

mcar.transform(input, na_prob)

Arguments

input

data to transform using MCAR mechanism

na_prob

desired probability of NA count defined as the probability of success in a Bernoulli trial.

Value

a data frame or matrix containing NAs

Examples

set.seed(123)
data <- data.frame(x1 = stats::rnorm(100),x2 = stats::rnorm(100),y = stats::rnorm(100))
mcar.transform(data,na_prob = .25)
#> Warning: One or more columns are at risk of being entirely NA.
#>               x1          x2           y
#> 1   -0.560475647          NA  2.19881035
#> 2   -0.230177489  0.25688371  1.31241298
#> 3    1.558708314 -0.24669188 -0.26514506
#> 4    0.070508391          NA  0.54319406
#> 5    0.129287735 -0.95161857 -0.41433995
#> 6             NA -0.04502772          NA
#> 7    0.460916206 -0.78490447 -0.78860284
#> 8             NA -1.66794194 -0.59461727
#> 9   -0.686852852 -0.38022652  1.65090747
#> 10  -0.445661970  0.91899661 -0.05402813
#> 11   1.224081797 -0.57534696  0.11924524
#> 12   0.359813827  0.60796432  0.24368743
#> 13   0.400771451 -1.61788271  1.23247588
#> 14            NA -0.05556197          NA
#> 15  -0.555841135  0.51940720          NA
#> 16            NA  0.30115336  1.67569693
#> 17            NA  0.10567619 -0.44116322
#> 18            NA -0.64070601 -0.72306597
#> 19            NA          NA -1.23627312
#> 20  -0.472791408 -1.02412879 -1.28471572
#> 21            NA  0.11764660 -0.57397348
#> 22            NA -0.94747461  0.61798582
#> 23            NA -0.49055744  1.10984814
#> 24  -0.728891229          NA          NA
#> 25  -0.625039268  1.84386201 -0.36365730
#> 26  -1.686693311          NA  0.05974994
#> 27   0.837787044          NA -0.70459646
#> 28            NA  0.07796085 -0.71721816
#> 29  -1.138136937          NA  0.88465050
#> 30   1.253814921 -0.07130809 -1.01559258
#> 31            NA  1.44455086  1.95529397
#> 32            NA          NA          NA
#> 33   0.895125661  0.04123292  0.21453883
#> 34   0.878133488          NA -0.73852770
#> 35   0.821581082 -2.05324722 -0.57438869
#> 36   0.688640254  1.13133721 -1.31701613
#> 37            NA          NA -0.18292539
#> 38            NA          NA  0.41898240
#> 39            NA  1.90910357  0.32430434
#> 40  -0.380471001 -1.44389316 -0.78153649
#> 41  -0.694706979          NA          NA
#> 42            NA -0.26219749 -0.50219872
#> 43            NA -1.57214416  1.49606067
#> 44   2.168955965 -1.51466765 -1.13730362
#> 45   1.207961998 -1.60153617 -0.17905159
#> 46  -1.123108583 -0.53090652  1.90236182
#> 47  -0.402884835          NA          NA
#> 48  -0.466655354  0.68791677 -1.35984070
#> 49   0.779965118  2.10010894 -0.66476944
#> 50  -0.083369066          NA  0.48545998
#> 51            NA  0.78773885 -0.37560287
#> 52            NA  0.76904224 -0.56187636
#> 53  -0.042870457  0.33220258          NA
#> 54   1.368602284 -1.00837661  0.09049665
#> 55  -0.225770986 -0.11945261  1.59850877
#> 56            NA -0.28039534 -0.08856511
#> 57            NA  0.56298953          NA
#> 58   0.584613750          NA  0.63075412
#> 59   0.123854244  0.97697339 -0.11363990
#> 60   0.215941569 -0.37458086          NA
#> 61            NA  1.05271147 -0.52111732
#> 62            NA -1.04917701 -0.48987045
#> 63  -0.333207384 -1.26015524  0.04715443
#> 64  -1.018575383  3.24103993  1.30019868
#> 65  -1.071791226 -0.41685759  2.29307897
#> 66   0.303528641  0.29822759  1.54758106
#> 67            NA  0.63656967 -0.13315096
#> 68   0.053004227 -0.48378063 -1.75652740
#> 69   0.922267468  0.51686204 -0.38877986
#> 70   2.050084686  0.36896453  0.08920722
#> 71  -0.491031166          NA  0.84501300
#> 72  -2.309168876          NA  0.96252797
#> 73   1.005738524 -0.03406725          NA
#> 74  -0.709200763          NA -1.39527435
#> 75  -0.688008616 -0.74133610          NA
#> 76   1.025571370 -1.09599627 -0.44655722
#> 77            NA  0.03778840  0.17480270
#> 78  -1.220717712  0.31048075  0.07455118
#> 79   0.181303480  0.43652348          NA
#> 80  -0.138891362          NA  0.02467498
#> 81   0.005764186 -1.06332613 -1.66747510
#> 82   0.385280401          NA          NA
#> 83            NA -0.34965039  0.38602657
#> 84            NA -0.86551286 -0.26565163
#> 85            NA          NA  0.11814451
#> 86   0.331781964 -0.19717589          NA
#> 87   1.096839013  1.10992029  0.22101947
#> 88   0.435181491  0.08473729  1.64084617
#> 89  -0.325931586  0.75405379 -0.21905038
#> 90   1.148807618          NA  0.16806538
#> 91   0.993503856          NA  1.16838387
#> 92   0.548396960 -0.32468591          NA
#> 93   0.238731735  0.09458353  1.14526311
#> 94  -0.627906076 -0.89536336 -0.57746800
#> 95   1.360652449          NA          NA
#> 96  -0.600259587  1.99721338          NA
#> 97   2.187332993  0.60070882  1.86685184
#> 98   1.532610626 -1.25127136 -1.35090269
#> 99  -0.235700359 -0.61116592  0.02098359
#> 100 -1.026420900 -1.18548008          NA