Transform data to missing values not at random

The MNAR mechanism assumes that the probability and cause of missingness are unknown to us and lie in the unobserved data. Under this mechanism we generate missing values by looking at each variable's lowest values and replace them with NA.

Usage

mnar.transform(input, target, features, na_rate)

Arguments

input

data to transform using MAR mechanism

target

variable to be used as causative feature

features

variables to which NA values are introduced using a causative feature

na_rate

proportion of missing values to be added to data

Value

a matrix or data frame containing NAs

Examples

set.seed(123)
data <- gen.mcar(100,rho = c(.15,.25,.12,.45,.34,.54),sigma = c(1,2,1,2),n_vars = 4, na_prob = 0)

mnar.transform(data,"V1",c("V2","V3"), na_rate = .25)
#>               V1          V2         V3          V4
#> 1   -0.560475647          NA 1.92166997 -0.10870782
#> 2   -0.230177489  0.43890141 1.22990170 -0.07693965
#> 3    1.558708314 -0.02018912 0.11332164 -0.34213763
#> 4    0.070508391          NA 0.51261534 -1.22758235
#> 5    0.129287735          NA         NA -1.40156838
#> 6    1.715064987  0.42548293         NA  1.60862474
#> 7    0.460916206          NA         NA -3.65506210
#> 8   -1.265061235          NA         NA -2.24347826
#> 9   -0.686852852          NA 1.38909629  2.38387824
#> 10  -0.445661970  1.68349957         NA  3.07054916
#> 11   1.224081797          NA 0.37304029  2.80692212
#> 12   0.359813827  1.31011577 0.37575598  1.98141531
#> 13   0.400771451          NA 1.15408996 -2.04864243
#> 14   0.110682716 -0.07666186         NA -1.25249611
#> 15  -0.555841135  0.86030892         NA -1.56643493
#> 16   1.786913137  1.13156617 2.08830791  4.21896599
#> 17   0.497850478  0.35831629         NA -0.01978737
#> 18  -1.966617157          NA         NA -4.58910274
#> 19   0.701355902          NA         NA  1.61466054
#> 20  -0.472791408          NA         NA -0.72096614
#> 21  -1.067823706 -0.08771603         NA -1.02676922
#> 22  -0.217974915          NA 0.46258112  1.59948550
#> 23  -1.026004448          NA 0.77317256 -2.24098233
#> 24  -0.728891229          NA 0.47897832  0.77256775
#> 25  -0.625039268  3.45848931         NA -0.62367973
#> 26  -1.686693311          NA         NA -0.81710759
#> 27   0.837787044  0.71678293         NA  0.20298897
#> 28   0.153373118  0.20016954         NA  0.51466709
#> 29  -1.138136937          NA 0.48857687  1.16051478
#> 30   1.253814921  0.23514186         NA  0.24816516
#> 31   0.426464221  2.98435368 2.11331690  4.32329398
#> 32  -0.295071483  0.80427002         NA -1.84825272
#> 33   0.895125661  0.35007052 0.43417554 -0.05785329
#> 34   0.878133488          NA         NA  2.18259134
#> 35   0.821581082          NA         NA -0.31655188
#> 36   0.688640254  2.44366660         NA -2.89312439
#> 37   0.553917654          NA         NA -2.47531237
#> 38  -0.061911711  1.44457797 0.45043497  0.40690741
#> 39  -0.305962664  3.68321913 0.39565082  2.32810853
#> 40  -0.380471001          NA         NA -1.94431513
#> 41  -0.694706979  1.17927660         NA  0.02242443
#> 42  -0.207917278          NA         NA  0.66403287
#> 43  -1.265396352          NA 0.99562998  1.71806274
#> 44   2.168955965          NA         NA  0.24607600
#> 45   1.207961998          NA         NA -0.02262170
#> 46  -1.123108583          NA 1.51002281 -2.29831263
#> 47  -0.402884835          NA         NA -1.10630322
#> 48  -0.466655354  1.22027075         NA -2.02608923
#> 49   0.779965118  4.38668614         NA -0.13691625
#> 50  -0.083369066          NA 0.34005846 -0.64315612
#> 51   0.253318514  1.63364829         NA  1.84672281
#> 52  -0.028546755  1.51211856         NA -3.01713255
#> 53  -0.042870457  0.64402694         NA -0.77516374
#> 54   1.368602284          NA 0.34530451  0.92410660
#> 55  -0.225770986 -0.30393353 1.47558074 -0.24809495
#> 56   1.516470604 -0.09950470 0.27028656  1.62930176
#> 57  -1.548752804  0.64861389 0.70237616  0.43073881
#> 58   0.584613750          NA 0.72352848  0.80102662
#> 59   0.123854244  1.96899609         NA -3.08848203
#> 60   0.215941569          NA         NA  2.50143679
#> 61   0.379639483  2.19549401         NA  0.18161627
#> 62  -0.502323453          NA         NA -0.47917608
#> 63  -0.333207384          NA         NA -0.55051359
#> 64  -1.018575383  6.10316898 1.27002998  3.47416178
#> 65  -1.071791226          NA 1.90927162  1.93673545
#> 66   0.303528641  0.68076548 1.59363150  1.42550096
#> 67   0.448209779  1.39319798 0.03672721  1.20133780
#> 68   0.053004227          NA         NA -3.08886788
#> 69   0.922267468  1.29870877         NA  2.05494510
#> 70   2.050084686  1.34460553 0.62936225  1.44212557
#> 71  -0.491031166          NA 0.67440628  3.71863157
#> 72  -2.309168876          NA 0.35665245 -3.67433945
#> 73   1.005738524  0.23435792 0.90870656  0.77477643
#> 74  -0.709200763  3.99598095         NA  0.58428732
#> 75  -0.688008616          NA 0.58574037  0.43838219
#> 76   1.025571370          NA         NA -0.92656278
#> 77  -0.284773007 -0.01071018 0.10058226 -1.56153920
#> 78  -1.220717712  0.24772062         NA -0.65152227
#> 79   0.181303480  0.91756035 0.49477946  0.74174666
#> 80  -0.138891362          NA         NA -3.00198988
#> 81   0.005764186          NA         NA -1.92834526
#> 82   0.385280401  2.61337111 0.91218148  1.94178380
#> 83  -0.370660032          NA 0.25053676  0.05555461
#> 84   0.644376549          NA         NA -1.67866638
#> 85  -0.220486562          NA 0.03912956  0.47385172
#> 86   0.331781964 -0.29035550 0.19579184  2.33906496
#> 87   1.096839013  2.52377700 0.58003103  2.45839981
#> 88   0.435181491  0.29811160 1.69869779  0.13672509
#> 89  -0.325931586  1.39326536         NA -0.70501919
#> 90   1.148807618          NA 0.40766212  1.41099113
#> 91   0.993503856  0.72208931 1.39334407  1.03626986
#> 92   0.548396960          NA 1.12691499 -1.98909747
#> 93   0.238731735  0.25864634 1.17234599  2.53389439
#> 94  -0.627906076          NA         NA -2.76846310
#> 95   1.360652449          NA 2.16246643  1.33402169
#> 96  -0.600259587  3.76915594 0.08093331  2.83854922
#> 97   2.187332993  1.84402470 2.39780535  2.72163176
#> 98   1.532610626          NA         NA  0.80617075
#> 99  -0.235700359          NA         NA -2.39399903
#> 100 -1.026420900          NA 0.85019544 -1.05282380