Stochastic regression imputation (SRI)

This method corrects the lack of variability in conditional mean imputation (CMI) by adding an error term to the conditional mean calculation. SRI is more effective than CMI in reducing bias in the imputed values and works well with MCAR and MAR data.

Usage

stoc.impute(
  data,
  family = "AUTO",
  tol = NULL,
  robust = FALSE,
  char_to_factor = FALSE,
  verbose = FALSE
)

Arguments

data

a numeric matrix or data frame of at least 2 columns.

family

the distribution family of your observations. The family arguments defaults to 'AUTO'; and it will automatically select a distribution family (gaussian, binomial, multinomial) based on the type of variable (numeric or factor). The distribution family dictates the regression model used (lm,glm, multinom). However, the user can change the family argument to match his response variable distribution and the function will adapt to this input by using the generalized linear model or beta regression.

tol

tolerance,a numeric vector of length 1 used as multiplicative factor to standard deviation for generalized linear models. As the sample size increases, the tolerance value should be decreased to represent the decreasing variability of the sample estimate.

robust

logical indicated whether to use robust estimation methods or ignore them. If set to 'TRUE', the function will make use of robust linear and generalized linear models to make its prediction.

char_to_factor

transform character variable to unordered factor variable

verbose

verbose error handling

Value

a matrix or data frame containing the imputed dataset.

Examples

data <- data.frame(x1 = c(stats::rnorm(87),rep(NA,13)),
x2 = stats::rnorm(100),y = stats::rnorm(100))
stoc.impute(data,tol = 1e-3)
#>               x1          x2            y
#> 1    1.712304977  1.48403093 -0.028395055
#> 2   -0.326143893 -1.18620659 -2.249051090
#> 3    0.373004656  0.10107915  0.031526005
#> 4   -0.227684065  0.53298929  0.205561208
#> 5    0.020450709  0.58673534 -0.155345347
#> 6    0.314057664 -0.30174666  0.568288621
#> 7    1.328214696  0.07950200  1.010677965
#> 8    0.121318377  0.96126415 -0.517982430
#> 9    0.712842320 -1.45646592 -0.294095327
#> 10   0.778860030 -0.78173971  0.397842209
#> 11   0.914773271  0.32040231 -0.550223741
#> 12  -0.574394552 -0.44478198  0.091267383
#> 13   1.626881214  1.37000399 -1.961707599
#> 14  -0.380956739  0.67325386 -1.119899721
#> 15  -0.105784168  0.07216675 -1.327755476
#> 16   1.404050268 -1.50775732 -0.853623697
#> 17   1.294083906  0.02610023 -0.693304528
#> 18  -1.089991872 -0.31641587  0.382305143
#> 19  -0.873071000 -0.10234651  0.982113003
#> 20  -1.358079059 -1.18155923 -0.727383529
#> 21   0.181847193  0.49865804 -0.996838982
#> 22   0.164840868 -1.03895644 -1.041688859
#> 23   0.364114687 -0.22622198 -0.414588726
#> 24   0.552157714  0.38142583 -0.239029068
#> 25  -0.601892846 -0.78351579  0.483617534
#> 26  -0.993698591  0.58299141 -0.321324842
#> 27   1.026785056 -1.31651040 -2.078489269
#> 28   0.751061303 -2.80977468 -0.091434276
#> 29  -1.509166537  0.46496799  1.187186811
#> 30  -0.095147451  0.84053983  1.191601269
#> 31  -0.895947823 -0.28584542 -0.788963218
#> 32  -2.070751071  0.50412625 -1.547776544
#> 33   0.150120131 -1.15591653  2.458060492
#> 34  -0.079211709 -0.12714861 -0.162421942
#> 35  -0.097369268 -1.94151838 -0.097451250
#> 36   0.216152542  1.18118089  0.420574189
#> 37   0.882465164  1.85991086 -1.614039457
#> 38   0.205597504  1.07401226 -0.728219111
#> 39  -0.616435843 -0.02734697 -1.540442405
#> 40  -0.734799252 -0.03333034 -0.693094614
#> 41  -0.131802793 -1.51606762  0.118849433
#> 42   0.310016987  0.79038534 -1.364709458
#> 43  -1.039680353 -0.21073418  0.589982679
#> 44  -0.184308869 -0.65674293  0.289344029
#> 45   0.967267260 -1.41202579 -0.904215026
#> 46  -0.108280091 -0.29976250  0.226324942
#> 47  -0.698420668 -0.84906114  0.748081162
#> 48  -0.275945168 -0.39703052  1.061095253
#> 49   1.114648545 -1.21759999 -0.212848279
#> 50   0.550043961  1.68758948 -0.093636794
#> 51   1.236675800 -0.01600253 -0.086714135
#> 52   0.139097857  1.07494508  1.441461756
#> 53   0.410275096 -2.60169967  1.125071892
#> 54  -0.558456912 -0.45319783  0.834401568
#> 55   0.605370669 -0.67548229 -0.287340800
#> 56  -0.506333542 -1.22292618  0.373241434
#> 57  -1.420565505  1.54660915  0.403290331
#> 58   0.127992966 -1.41528192 -1.041673294
#> 59   1.945851218  0.31839026 -1.728304515
#> 60   0.800914340  0.84643629  0.641830028
#> 61   1.165253390  0.17819019 -1.529310531
#> 62   0.358855723 -0.87525548  0.001683688
#> 63  -0.608557178  0.94116581  0.250247821
#> 64  -0.202240855  0.17058808  0.563867390
#> 65  -0.273248107 -1.06349791  0.189426238
#> 66  -0.468699780 -1.38804905 -0.732853806
#> 67   0.704167284  2.08671743  0.986365860
#> 68  -1.197363502 -0.67850315  1.738633767
#> 69   0.866366132 -1.85557165  0.881178809
#> 70   0.864152486  0.53325936 -1.943650901
#> 71  -1.198622359  0.31023026  1.399576185
#> 72   0.639491998 -1.35383434 -0.056055946
#> 73   2.430226652 -1.94295641  0.524914279
#> 74  -0.557215482 -0.11630252  0.622033236
#> 75   0.844904241  1.13939629 -0.096686073
#> 76  -0.782201846  0.63612404 -0.075263198
#> 77   1.110711418 -0.49293742  1.019157069
#> 78   0.249824720 -0.83418823  0.711601922
#> 79   1.651915392  0.27106676  0.990262246
#> 80  -1.458970729  0.15735335  2.382926695
#> 81  -0.051297886  0.62971175  0.664415864
#> 82  -0.526925176 -0.39579795  0.207381157
#> 83  -0.197264869  0.89935405 -2.210633111
#> 84  -0.629578743 -0.83081153  2.691714003
#> 85  -0.833843581 -0.33054470 -0.482676822
#> 86   0.578722375  0.74081452  2.374734715
#> 87  -1.087580714  0.98997161  0.374643568
#> 88  -1.548420385 -1.93850470  1.538430199
#> 89   0.407891406  0.10719041 -0.109710321
#> 90   0.794801562  0.60877901  0.511470755
#> 91   1.242187286 -1.45082431  0.213957980
#> 92  -0.121026773  0.48062560 -0.186120699
#> 93  -0.136362637 -0.82817427 -0.120393825
#> 94   1.138510265  1.02025301  1.012834336
#> 95  -1.383250457  0.53848203 -0.201458147
#> 96  -0.008374891  0.76905229 -2.037682494
#> 97   0.499296579  0.12071933 -0.195889249
#> 98  -1.485782488  0.86364843  0.539790606
#> 99   0.127811884  1.38051453  0.616455716
#> 100  1.488718212  1.96624802  0.616567817