Calculate the geometric standardized score

Compute the geometric z score of each strictly positive value in the sample, relative to the geometric mean and standard deviation typically used in log-normally distributed populations.

Usage

geom.zscore(x, ..., na.rm = FALSE)

Arguments

x

a numeric vector, matrix or data frame

...

additional parameters to be passed to 'geom.mean()'

na.rm

indicate whether or not to remove NAs

Value

a matrix or array of z-scores of length > 1L.

Details

The formula here makes use of the change of base property of logarithms.

Examples

d <- rlnorm(60,1,1.4)
geom.zscore(d)
#>  [1] -1.383639914 -0.308561264 -0.241446981  0.819578374 -1.172421586
#>  [6] -0.626914902  1.203299597  1.087485196  0.912627420  1.358433139
#> [11]  0.054506910 -0.648600941  0.368679125  0.819145597  2.031297503
#> [16] -1.885971305  0.116419973  0.457776428 -0.176872570 -1.521711082
#> [21]  0.535347046 -0.825025073  0.324535473  0.683902145  1.020648024
#> [26] -1.521748253  0.839738983 -0.805249687 -1.281230341  1.047812454
#> [31]  0.897459245 -0.345050443  0.551486120 -0.897705333 -0.263144972
#> [36] -0.211874841 -0.379186892  0.528724808  1.247322512 -0.535776000
#> [41] -0.157244066  1.336276041 -0.422399544 -2.457693906  0.331801967
#> [46] -0.921093158 -1.021355410  1.143087782  2.062041445 -1.058253914
#> [51]  0.794366008 -0.212254708 -0.001160515  0.965430790  0.063134371
#> [56] -0.889689044 -0.548871519  0.485814933 -2.217850089  0.851818841

x <- matrix(rlnorm(60,1,1.4), ncol = 4)
mapply(geom.zscore, as.data.frame(x))
#>               V1           V2         V3         V4
#>  [1,] -1.6989624  0.867584410 -0.5263654 -1.4058539
#>  [2,]  0.1188023 -1.924556602 -0.4387176 -0.8729480
#>  [3,] -0.1240382 -1.543854643  0.9965401  0.1463040
#>  [4,]  1.5728694  0.940535900 -0.5247093  1.1533292
#>  [5,]  0.3097417 -0.766191755 -0.2067384 -0.8704847
#>  [6,]  0.2103243  0.894027454  0.1092892 -0.2962724
#>  [7,]  0.8923192 -0.244742875 -0.6069344  1.6063547
#>  [8,] -0.8523145  0.235963323  1.5104990 -1.2533980
#>  [9,] -0.1303753  0.546878988 -1.4897069  0.9455880
#> [10,]  1.4407649 -0.634260639  1.5567562  0.4167803
#> [11,]  0.8755470 -0.354704259 -0.5359645 -1.2305802
#> [12,] -1.3653374 -0.443829646  0.9773788 -0.7664968
#> [13,]  0.8234794  2.162179592 -1.9951707  0.7712531
#> [14,] -1.4821224 -0.007305883  0.6221932  0.2050469
#> [15,] -0.5906981  0.272276635  0.5516507  1.4513779