Stochastic regression imputation (SRI)

This method corrects the lack of variability in conditional mean imputation (CMI) by adding an error term to the conditional mean calculation. SRI is more effective than CMI in reducing bias in the imputed values and works well with MCAR and MAR data.

Usage

stoc.impute(
  data,
  family = "AUTO",
  tol = NULL,
  robust = FALSE,
  char_to_factor = FALSE,
  verbose = FALSE
)

Arguments

data

a numeric matrix or data frame of at least 2 columns.

family

the distribution family of your observations. The family arguments defaults to 'AUTO'; and it will automatically select a distribution family (gaussian, binomial, multinomial) based on the type of variable (numeric or factor). The distribution family dictates the regression model used (lm,glm, multinom). However, the user can change the family argument to match his response variable distribution and the function will adapt to this input by using the generalized linear model or beta regression.

tol

tolerance,a numeric vector of length 1 used as multiplicative factor to standard deviation for generalized linear models. As the sample size increases, the tolerance value should be decreased to represent the decreasing variability of the sample estimate.

robust

logical indicated whether to use robust estimation methods or ignore them. If set to 'TRUE', the function will make use of robust linear and generalized linear models to make its prediction.

char_to_factor

transform character variable to unordered factor variable

verbose

verbose error handling

Value

a matrix or data frame containing the imputed dataset.

Examples

data <- data.frame(x1 = c(stats::rnorm(87),rep(NA,13)),
x2 = stats::rnorm(100),y = stats::rnorm(100))
sri_data <- stoc.impute(data,tol = 1e-3)
summary(sri_data)
#>        x1                 x2                  y           
#>  Min.   :-2.07075   Min.   :-2.809775   Min.   :-2.24905  
#>  1st Qu.:-0.56244   1st Qu.:-0.831656   1st Qu.:-0.58594  
#>  Median : 0.12457   Median : 0.005049   Median : 0.01660  
#>  Mean   : 0.08157   Mean   :-0.089453   Mean   : 0.02355  
#>  3rd Qu.: 0.75801   3rd Qu.: 0.631315   3rd Qu.: 0.61793  
#>  Max.   : 2.43023   Max.   : 2.086717   Max.   : 2.69171